test(supabase): align pgtap suite with multi-leg trade routes
Sweep updates the pgtap suite plus seed to match recent schema and rpc
changes that were missed when their migrations landed.

Trade routes (migration 20260604000011): split resource_id and
quantity_per_transition off public.trade_routes onto trade_route_legs.
Adjust every test fixture and the local seed to insert routes without
those columns and emit one send-leg row per route. Update
propose_trade_route and replace_trade_route call sites to the new jsonb
legs signature introduced in 20260604000012/013. Drop now-meaningless
rls cases that exercised the removed quantity check and same-world
resource trigger.

Snapshot rls (migration 20260604000004): world admin direct delete no
longer raises 42501 because the policy was dropped; rls now silently
filters. Switch those assertions from throws_ok to a row-count check.
Reset to postgres before the duplicate-insert path in
apply_turn_transition_settlement_snapshots_test now that authenticated
has insert revoked.

Other targeted fixes: drop the advance_world_turn_if_current
assertions from local_seed_topology_test (rpc deleted in
20260603000010); replace the rpc-driven notification seed in
notifications_rls_test with direct inserts; expect readinessReset=2
in apply_turn_transition_full_test to match the unconditional update
in 20260603000009 and rewind the stockpile before the idempotent
retry so the 20260604000003 revalidation does not reject it; switch
the construction-job case in bulk_standard_job_assignment_test to
lives_ok after 20260604000008 merged construction into the same rpc,
and populate death_cause_category on dead-citizen updates so the
20260602000008 check constraint passes; cast 0 to numeric for the
numeric current_count compare; reorder world_admins after worlds in
set_construction_project_workers_test to fix the fk; correct three
plan counts that were off by one (start_turn_transition,
set_construction_project_workers, propose_trade_route).

+++ supabase/tests/remove_managed_population_instance_test.sql
@@ -312,7 +312,7 @@ select
       where
         mpi.id = 'fb800000-0000-0000-0000-000000000002'
     ),
-    0,
+    0::numeric,
     'managed population instance current_count is 0 after successful removal'
   );
 
@@ -351,7 +351,7 @@ select
       where
         mpi.id = 'fb800000-0000-0000-0000-000000000001'
     ),
-    0,
+    0::numeric,
     'managed population instance current_count is 0 after admin success'
   );
